enabled & updated phpstan

+++ api/server.php
@@ -35,7 +35,7 @@ try
             {
                 $userid = isset($_POST['userid']) ? (int)$_POST['userid'] : 0;
                 $token = isset($_POST['token']) ? $_POST['token'] : "";
-                $address = isset($_POST['address']) ? utf8_encode($_POST['address']) : null;
+                $address = isset($_POST['address']) ? (int)utf8_encode($_POST['address']) : null;
                 $port = isset($_POST['port']) ? (int)utf8_encode($_POST['port']) : null;
                 $private_port = isset($_POST['private_port']) ? (int)utf8_encode($_POST['private_port']) : null;
                 $server_name = isset($_POST['name']) ? $_POST['name'] : "";
